Fixes #24
When implementation classes derive from interface classes, methods
which may actually be missing in the implementation are substituted
from the interface class. Thus, it is necessary to prevent the
implementations from inheriting from interfaces. This commit
enforces verification of this condition within the _implements_
decorator.

The README has also been updated to inform the user of this pattern