Implements version 0.3.0 (#26)
* Fixes #25

Data-descriptors are enforced on the implementation only when the
interface implements it. Unittest added to verify this case.

* #22: Added unittests to verify signatures for data-descriptors

* Fixes #22: Added code to verify data-descriptor signatures in addition to presence

* #22: Removed redundant parantheses in class name

* Fixes #24

When implementation classes derive from interface classes, methods
which may actually be missing in the implementation are substituted
from the interface class. Thus, it is necessary to prevent the
implementations from inheriting from interfaces. This commit
enforces verification of this condition within the _implements_
decorator.

The README has also been updated to inform the user of this pattern

* Fixed flake8 issue. Extra blank lines.

* Updated CHANGELOG for version 0.3.0

* Retaining only Kamil's name for package author in setup.py. Multiple entries don't render correctly on pypi